+++ src/main/java/paperJ/cutservice/domain/paper/PaperColor.java
@@ -15,7 +15,7 @@ public class PaperColor {
     private String color;
 
     @Column(name = "PRICE_PER_SHEET")
-    private double pricePerSheet;
+    private int pricePerSheet;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "PAPER_TYPE_ID")
@@ -25,7 +25,7 @@ public class PaperColor {
     protected PaperColor() {}
 
     // 사용자에게 입력받는 초기 정보로 생성
-    public PaperColor(String color, double pricePerSheet, PaperType paperType) {
+    public PaperColor(String color, int pricePerSheet, PaperType paperType) {
         this.color = color;
         this.pricePerSheet = pricePerSheet;
         this.paperType = paperType;
